+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/tool/codenarc/CodenarcTool.groovy
@@ -52,6 +52,14 @@ class CodenarcTool implements QualityTool {
     }
 
     @Override
+    String getToolInfo(Project project, QualityExtension extension, List<ProjectSources> langs) {
+        if (langs.contains(ProjectSources.Groovy)) {
+            return 'CodeNarc: ' + (extension.codenarc.get() ? extension.codenarcVersion.get() : 'disabled')
+        }
+        return null
+    }
+
+    @Override
     Reporter createReporter(Object param, Provider<ConfigsService> configs) {
         return new CodeNarcReporter(param as Properties)
     }